+++ extension/src/webcat/hooks/entry-page.ts
@@ -1,17 +1,17 @@
-import {
-  eventHook,
-  serviceWorkerHook,
-  sharedWorkerHook,
-  wasmHook,
-  workerHook,
-} from "./core";
+// This file is to be compiled, minified and embedded as a string to be
+// dynamically updated with data and injected into scripts
+// by response.ts
+
+import { eventHook } from "./events";
+import { wasmHook } from "./wasm";
+import { serviceWorkerHook, sharedWorkerHook, workerHook } from "./workers";
 
 console.log("[WEBCAT] Installing page hook");
-
+const data = "__DATA_PLACEHOLDER__";
 const scope = {};
 
-wasmHook(scope);
-sharedWorkerHook(scope);
-serviceWorkerHook(scope);
-workerHook(scope);
-eventHook(scope);
+wasmHook(scope, data);
+sharedWorkerHook(scope, data);
+serviceWorkerHook(scope, data);
+workerHook(scope, data);
+eventHook(scope, data);
